Extract "Copy transcript" button into a separate component
Extract this out of VideoPlayerApp since we expect to add some extra complexity
here soon.
